+++ src/DaisyDuino.h
@@ -1,33 +1,22 @@
-#ifndef DSY_ARDUINO_AUDIO_H
-#define DSY_ARDUINO_AUDIO_H
+#ifndef DSY_DSYDUINO_H
+#define DSY_DSYDUINO_H
 
 #include <stdio.h>
 #include "Arduino.h"
 #include "DaisyDSP.h"
 #include "daisy_pod.h"
 #include "daisy_patch.h"
-#include "utility/audio.h"
-#include "utility/sai.h"
-#include "utility/sdram.h"
+#include "AudioClass.h"
 
 using namespace daisy;
 
-enum DaisyDuinoDevice {
+enum DaisyDuinoDevice : short{
     DAISY_SEED,
     DAISY_POD,
     DAISY_PETAL,
     DAISY_FIELD,
     DAISY_PATCH,
     DAISY_LAST,
-};
-
-enum DaisyDuinoSampleRate {
-	AUDIO_SR_8K,
-    AUDIO_SR_16K,
-    AUDIO_SR_32K,
-    AUDIO_SR_48K,
-    AUDIO_SR_96K,
-    AUDIO_SR_LAST,
 };
 
 //TODO
@@ -184,9 +173,13 @@ class GateIn
         Switch sw;
 };
 
+class AudioClass; ///forward declaration
+
 class DaisyHardware
 {
 public:
+	DaisyHardware() {}
+
     Switch buttons[2];
     Encoder encoder;
     Led leds[2];
@@ -195,124 +188,12 @@ public:
     int num_channels;
     int numSwitches, numLeds, numGates;
 
-    void Init(float control_update_rate, DaisyDuinoDevice device)
-    {
-	device_ = device;
-	num_channels = 2;
-	numGates = 0;
-	
-	switch (device){
-	    case DAISY_SEED:
-		break;
-		
-	    case DAISY_POD:
-	        buttons[0].Init(control_update_rate, true, PIN_POD_SWITCH_1, INPUT_PULLUP);
-	        buttons[1].Init(control_update_rate, true, PIN_POD_SWITCH_2, INPUT_PULLUP);
-				
-			encoder.Init(control_update_rate, PIN_POD_ENC_A, PIN_POD_ENC_B, PIN_POD_ENC_CLICK, INPUT_PULLUP, INPUT_PULLUP, INPUT_PULLUP);
+    void Init(float control_update_rate, DaisyDuinoDevice device);
 
-			leds[0].Init(PIN_POD_LED_1_RED, PIN_POD_LED_1_GREEN, PIN_POD_LED_1_BLUE);    
-	        leds[1].Init(PIN_POD_LED_2_RED, PIN_POD_LED_2_GREEN, PIN_POD_LED_2_BLUE);
-
-			numSwitches = numLeds = 2;
-			break;
-
-	    case DAISY_PETAL:
-		break;
-	    
-		case DAISY_FIELD:
-		break;
-	    
-		case DAISY_PATCH:
-			encoder.Init(control_update_rate, PIN_PATCH_ENC_A, PIN_PATCH_ENC_B, PIN_PATCH_ENC_CLICK, INPUT_PULLUP, INPUT_PULLUP, INPUT_PULLUP);
-			gateIns[0].Init(20, INPUT, true);
-			gateIns[1].Init(19, INPUT, true);
-			numSwitches = numLeds = 0;
-			numGates = 2;
-			num_channels = 4;
-		break;
-	    default:
-		break;
-	}
-    }
-
-    void DebounceControls()
-    {
-	for (int i = 0; i < numSwitches; i++)
-	{
-	    buttons[i].Debounce();
-	}
-	
-		encoder.Debounce();
-		for (int i = 0; i < numGates; i++)
-		{
-			gateIns[i].Debounce();
-		}
-	}
+    void DebounceControls();
 
 private:
     DaisyDuinoDevice device_;
 };
-
-class AudioClass
-{
-    public:
-        AudioClass() {}
-
-        // Initializes the audio for the given platform, and returns a DaisyHardware object
-		//default samplerate is 48kHz
-        DaisyHardware init(DaisyDuinoDevice device, DaisyDuinoSampleRate sr = AUDIO_SR_48K);
-        
-		/** for bwd compatibility */
-		void begin(AudioHandle::AudioCallback cb);
-
-		/** for bwd compatibility. Does not work with Daisy Patch!! */
-		void begin(AudioHandle::InterleavingAudioCallback cb);
-
-		/** for bwd compatibility */			
-		void end();
-
-		/** for bwd compatibility */			
-		float get_samplerate();
-			
-		/** for bwd compatibility */			
-		float get_callbackrate();
-		
-		/** for bwd compatibility */			
-		inline float get_blocksize();
-		
-		/** Interleaving callback not supported on patch. */
-		void StartAudio(AudioHandle::InterleavingAudioCallback cb);
-
-		void StartAudio(AudioHandle::AudioCallback cb);
-
-		/** Won't work on Patch! */
-		void ChangeAudioCallback(AudioHandle::InterleavingAudioCallback cb);
-
-		void ChangeAudioCallback(AudioHandle::AudioCallback cb);
-
-		void StopAudio();
-
-		/** Works with the new samplerates */
-		void SetAudioSampleRate(SaiHandle::Config::SampleRate samplerate);
-
-		float AudioSampleRate();
-
-		void SetAudioBlockSize(size_t blocksize);
-
-		size_t AudioBlockSize();
-
-		float AudioCallbackRate();
-				
-    private:
-		float callback_rate_;
-        
-		AudioHandle audio_handle;
-		DaisyDuinoDevice _device;
-        dsy_sdram_handle sdram_handle;
-	void ConfigureSdram();
-};
-
-extern AudioClass DAISY;
 
 #endif
